refactor(TKT-699): refactor: extract repoWorktrees detection into shared context.ts module
